Changed examples corner and removed datasets API for now...
