+++ dnet/__init__.py
@@ -1,4 +1,3 @@
-from . import datasets
 from . import layers
 from . import losses
 from . import models
