+++ dnet/data/__init__.py
@@ -8,8 +8,3 @@ class Dataloader:
     def __init__(self, train_data: Tuple[Tensor, Tensor], val_data: Tuple[Tensor, Tensor]):
         self.train_data = train_data
         self.val_data = val_data
-
-
-__all__ = [
-    "Dataloader"
-]
